% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/coocA.R
\name{coocA}
\alias{coocA}
\title{Wrapper function for computing coocurrences
The function parses user input, checks for existing corpora, formulates query phrase}
\usage{
coocA(corp, piv, poscooc = "QLF SUB VBE", dis = 5, coeff = "dice",
  selec = 1, nbs = 30, D = 0, F = Inf, attr = "", val = "",
  corp_lemma = "lemma", corp_pos = "pos")
}
\arguments{
\item{corp}{Corpus name}

\item{piv}{Searched word}

\item{poscooc}{POS of the coocurences}

\item{dis}{Coocurence window}

\item{coeff}{Measure}

\item{selec}{???}

\item{nbs}{Number of coocurrences}

\item{D}{Start (index of corpus position)}

\item{F}{(index of corpus position)}

\item{attr}{Structural attribute to filter the results}

\item{val}{Value of the structural attribute}

\item{corp_lemma}{Name of the positional attribute for lemma}

\item{corp_pos}{Name of the positional attribute for POS}
}
\value{
'cooca' object
}
\description{
Wrapper function for computing coocurrences
The function parses user input, checks for existing corpora, formulates query phrase
}
\examples{
cooc <- coocA(corp="PATROLOGIA",piv="hortus",dis=3)
}

